Completes scaffholding now working on controller.
This part completes the scaffholding to make this work.

Now working on the controller.
Currently the controller just trys to get the max speed of the
circut. Once the car comes off it then just keeps the car
going at that max speed. However, this does have some problems
and could probably do with multiple attempts to get an average.