Add images for docs and protocol for hardware.
This adds some screen shots that will be included in the documentation
at the end of the project. These have been taken now to show the spike work
that was undertaken.

The protocol is an initial idea how to do the controlling. I'll look for
some feed back from the Peter who built the hardware.